refactor: introduce world-chain-proof-core and consolidate proof crates
Extract a shared `world-chain-proof-core` crate with zero SP1/vsock dependencies
containing all types used by both the Succinct and Nitro TEE backends: boot values,
range hardfork config, proof claim/input/public-values, witness data, preimage store,
blob oracle, and artifacts.

- Add `world-chain-proof-core` with `RollupConfigHashError`, `hash_rollup_config`,
  `hash_world_rollup_config_generic`, shared range/boot/witness/artifact types
- Add `world-chain-proof` wrapper crate with `sp1` and `aws_nitro` feature flags
- Delete `world-chain-proof-client` and `world-chain-proof-host` (types absorbed into
  core and `succinct-host-utils` respectively; no duplicate type systems)
- Move `WorldProverIdentity`, `WorldFaultDisputeGameConfig`, `WorldFaultProofServiceConfig`,
  `RangeSplitCount` into `succinct-host-utils`; gate on `WorldRangeHardforkConfig` from core
- Strip backward-compat re-exports from `succinct-client-utils`; all consumers import
  shared types directly from `world-chain-proof-core`
- Remove conversion boilerplate in `succinct-host-utils` (`range_proof_input`,
  `range_public_values`, `range_hardfork`, `range_spec_id`); `range_request` now
  takes `WorldRangeProofInput` directly
- Delete dead files left in `succinct-client-utils/src/` after the refactor
- Gate `aws_nitro` feature (tokio-vsock, NSM) in `world-chain-proof-nitro`
