--- crates/proof/client/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "world-chain-proof-client"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-alloy-primitives = { workspace = true, features = ["serde"] }
-serde.workspace = true
-thiserror.workspace = true
-world-chain-proof-protocol.workspace = true
-
-[dev-dependencies]
-serde_json.workspace = true
