+++ crates/proof/core/Cargo.toml
@@ -1,43 +1,28 @@
 [package]
-name = "world-chain-proof-succinct-client-utils"
+name = "world-chain-proof-core"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
+description = "Shared proof primitives for World Chain fault-proof backends (SP1 and Nitro TEE)."
 
 [dependencies]
 alloy-consensus.workspace = true
 alloy-eips.workspace = true
 alloy-primitives = { workspace = true, features = ["serde"] }
-alloy-rlp.workspace = true
 alloy-sol-types.workspace = true
-alloy-evm.workspace = true
-alloy-op-evm.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
-cfg-if = "1"
 kona-derive.workspace = true
-kona-driver.workspace = true
-kona-executor.workspace = true
 kona-genesis.workspace = true
 kona-preimage.workspace = true
 kona-proof.workspace = true
 kona-protocol.workspace = true
 kzg-rs.workspace = true
-op-alloy-consensus.workspace = true
-op-revm.workspace = true
-revm = { version = "38.0.0", default-features = false }
-revm-precompile = { version = "34.0.0", default-features = false }
 rkyv.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 sha2.workspace = true
-spin.workspace = true
 thiserror.workspace = true
-tracing.workspace = true
-world-chain-proof-core.workspace = true
-
-[dev-dependencies]
-serde_json.workspace = true
