+++ crates/proof/core/src/types.rs
@@ -58,12 +58,11 @@ mod tests {
             0x191a_1b1c,
             0x1d1e_1f20,
         ];
-
         assert_eq!(
             u32_to_u8(words),
             [
-                1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,
-                24, 25, 26, 27, 28, 29, 30, 31, 32,
+                1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22,
+                23, 24, 25, 26, 27, 28, 29, 30, 31, 32,
             ]
         );
     }
