+++ crates/proof/core/src/witness/preimage_store.rs
