+++ crates/proof/succinct/programs/Cargo.toml
@@ -29,6 +29,7 @@ sp1-zkvm = { version = "=6.1.0", features = ["verify"] }
 thiserror = "2"
 tracing = { version = "0.1", default-features = false }
 tracing-subscriber = { version = "0.3.18", features = ["fmt"] }
+world-chain-proof-core = { path = "../../core", default-features = false }
 world-chain-proof-succinct-client-utils = { path = "../utils/client", default-features = false }
 world-chain-proof-succinct-ethereum-client-utils = { path = "../utils/ethereum/client", default-features = false }
 world-chain-proof-succinct-range-utils = { path = "range/utils", default-features = false }
