+++ crates/proof/succinct/programs/aggregation/Cargo.toml
@@ -15,4 +15,5 @@ serde_cbor.workspace = true
 sha2.workspace = true
 sp1-lib.workspace = true
 sp1-zkvm.workspace = true
+world-chain-proof-core.workspace = true
 world-chain-proof-succinct-client-utils.workspace = true
