+++ crates/proof/succinct/programs/range/ethereum/Cargo.toml
@@ -10,6 +10,7 @@ repository.workspace = true
 kona-proof.workspace = true
 rkyv.workspace = true
 sp1-zkvm.workspace = true
+world-chain-proof-core.workspace = true
 world-chain-proof-succinct-client-utils.workspace = true
 world-chain-proof-succinct-ethereum-client-utils.workspace = true
 world-chain-proof-succinct-range-utils.workspace = true
