+++ crates/proof/succinct/utils/host/Cargo.toml
@@ -21,8 +21,7 @@ kona-protocol.workspace = true
 kzg-rs.workspace = true
 serde.workspace = true
 thiserror.workspace = true
-world-chain-proof-client.workspace = true
-world-chain-proof-host.workspace = true
-world-chain-proof-protocol.workspace = true
+serde_json.workspace = true
+world-chain-proof-core.workspace = true
 world-chain-proof-succinct-client-utils.workspace = true
 world-chain-proof-succinct-proof-utils.workspace = true
